SVSM: move heap_start declaration to other layout related symbols
In svsm.rs, there are a couple of declarations referring to the layout related
symbols defined in the linker script: {s,e}text, {s,e}data and so on. These
all are declared next to each other in svsm.rs, except for heap_start. Move
heap_start over to the rest for better code readability.

Signed-off-by: Nicolai Stange <nstange@suse.de>