SVSM/mm/alloc: factor out slab's slot deallocation logic
The SLAB_PAGE_SLAB, being implemented as an instance of a regular Slab, can
recurse when it needs to allocate a SlabPage for itself. A future commit will
address this by introducing a specialized implementation, SlabPageSlab, managing
the slab growing differently.

In order to faciliate code sharing between the Slab and SlabPageSlab
implementations, split out the slab slot deallocation logic in
Slab::deallocate() into a separate, internal function, ::deallocate_slot().

Signed-off-by: Nicolai Stange <nstange@suse.de>