SVSM/mm/alloc: make SlabPage's Slab backref optional
The SLAB_PAGE_SLAB, being implemented as an instance of a regular Slab, can
recurse when it needs to allocate a SlabPage for itself. A future commit will
address this by introducing a specialized implementation, SlabPageSlab, managing
the slab growing differently.

For pages associated with some Slab, there is a backreference from the Page
back to the owning Slab instance in order to enable the public
GlobalAlloc::dealloc() interface to find the respective owning Slab to forward
the deallocation request to. For pages wich will be associated with the
to-be-introduced SlabPageSlab, there will be no such Slab instance, but
fortunately objects allocated thereof, i.e. SlabPages, won't ever get freed
through that public interface anyway. Thus, a viable approach is to
simply have pages associated with a SlabPageSlab to not reference back to
anything. Prepare for this by making the Page::SlabPage() Slab backreference
optional.

Signed-off-by: Nicolai Stange <nstange@suse.de>