SVSM/mm/alloc: split out reusable functionality from Slab
The SLAB_PAGE_SLAB, being implemented as an instance of a regular Slab, can
recurse when it needs to allocate a SlabPage for itself. A future commit will
address this by introducing a specialized implementation, SlabPageSlab, managing
the slab growing differently.

In order to faciliate code sharing between the Slab and SlabPageSlab
implementations, split out the reusable functionality from struct Slab into
a new, internal struct SlabCommon to be shared between the two.

Signed-off-by: Nicolai Stange <nstange@suse.de>