SVSM/mm/alloc: minimize unsafe{} region in Slab::grow_slab()
There is no need for the whole Slab::grow_slab() function to be marked as
unsafe{} -- the only unsafe operation is the case of the memory returned
by SLAB_PAGE_SLAB to a PageSlab. Only wrap this single cast as unsafe{}.

Also, while touching the code, make the SlabPage allocation more robust by
not unwrap()ping the result, which could panic, but propagating Err()s, if any,
upwards the callchain instead.

Signed-off-by: Nicolai Stange <nstange@suse.de>