Stage2+SVSM: use common println! implementation on top of the log crate
Currently, the stage2 and svsm bin crates define their variant of the println!()
macro each, which gets used all over the place from the shared code. This
hinders making a lib crate out of that shared code.

Reimplement the println!() macro on the base of the common logging facility at
a central place and drop the individual definitions from the bin crates.

Signed-off-by: Nicolai Stange <nstange@suse.de>