sev: split sev_status_init() into init and verify stages
Currently, stage2 calls sev_status_init() first, so that the global SEV_FLAGS
is initialized when the subsequent sev_es_enabled() from stage2 attempts to
examine is.

The sev_es_enabed() check's purpose is to report an error on the serial via
emulated io instructions when SEV-ES is not enabled, because in this case
the SVSMIOPort console, utilizing the GHCB protocol, would not be available.

However, the early sev_status_init() would check the SEV status by itself and
panic on a not yet functional console if an unsupported combination, like SEV-ES
not being active in particular, is encountered. With that the fallback to the
emulated serial if !sev_es_enabled() is effectively dead code and the
unsupported configuration error report wouldn't reach the user at all.

Fix this by splitting sev_status_init() into two stages:
- the first one merely initializing SEV_FLAGS so that it's available to
  sev_es_enabled() and
- the second one, now moved into the new sev_status_verify(), running
  the extended status flags verification.

Invoke the latter sev_status_verify() after the console initialization from
stage2, so that its problem reports, if any, would reach the user.

Signed-off-by: Nicolai Stange <nstange@suse.de>