Stage2: initialize console before panicking for !SEV-ES
When the check for sev_es_enabled() in stage2 fails, the latter would attempt
to report an error on the console and panic. However, in this error path, the
console is not initialized and the message will get silently dropped. Make
stage2.rs to initialize the console in the error path if SEV-ES is not enabled.

Signed-off-by: Nicolai Stange <nstange@suse.de>