Stage2: check for SEV-ES before initializing paging
Currently, one of the first things stage2.rs does is to initialize the paging.
paging_init() attempts to obtain the C-bit from the cpuid page and would panic
if not found, at a stage where the console is not yet functional. As the cpuid
page is only available under SEV-SNP, this effectively makes the subsequent SEV
status checks from stage2.rs dead code.

A working paging subsystem is first needed when establishing a GHCB, for use
from the SVSMIOPort console among others.

So give the SEV-ES check in stage2.rs a chance to actually run and report a
meaningful error, if needed, by moving the page_init() invocation to after it.

Signed-off-by: Nicolai Stange <nstange@suse.de>