sev/ghcb: make GHCB initialization succeeed on !SEV_SNP
If SEV-ES is enabled, but SEV-SNP is not, the stage2 will attempt to report an
error through the SVSMIOport console. However, the latter relies on a functional
GHCB.

As it currently stands, the GHCB initialization will unconditionally issue a
pvalidate instruction and a subsequent invalidate_page HV request for mapping
the GCHB page as shared. These SEV-SNP specific operations would not work when
SEV-SNP was off, effectively rendering the error reporting to SVSMIOport
disfunctional in SEV-ES-only mode.

Fix this by making the SEV-SNP specific operations in the course of GHCB
initialization conditional on SEV-SNP status.

Signed-off-by: Nicolai Stange <nstange@suse.de>