Stage2/boot: provide IDT support for the 32 bit stage
A subsequent commit will make the boot code in boot_stage2.rs to proceed even
if launched under a non-SEV-SNP environment, with the purpose of eventually
reporting a meaningful error message once the console becomes functional.

This will e.g. involve accesses to some SEV related MSRs, which might not exist
and cause #GPs. In order to proceed gracefully, a #GP handler will be needed for
fixing up the fault.

Provide the necessary infrastructure, i.e. define and load an IDT for the 32 bit
stage and implement idt32_install_handler() for installing exception handlers at
it.

Signed-off-by: Nicolai Stange <nstange@suse.de>