Stage2/boot: implement __rdmsr_safe()/__wrmsr_safe() for 32 bit stage
A subsequent commit will make the boot code in boot_stage2.rs to proceed even
if launched under a non-SEV-SNP environment, with the purpose of eventually
reporting a meaningful error message once the console becomes functional.

This will involve rdmsr/wrmsr on the SEV related MSRs, which might not exist
and cause #GPs. Implement __rdmsr_safe() and __wrmsr_safe(), which are wrappers
around the rdmsr and wrmsr instructions respectively, but with #GPs fixed up
in cooperation with the new gp_msr_fixup_handler32() installed at the 32 bit
IDT.

Signed-off-by: Nicolai Stange <nstange@suse.de>