Stage2/boot: implement __vmgexit_safe() for 32 bit stage
A subsequent commit will make the boot code in boot_stage2.rs to proceed even
if launched under a non-SEV-SNP environment, with the purpose of eventually
reporting a meaningful error message once the console becomes functional.

This will involve an invocation of the GCHB MSR protocol, and, ultimately, a
vmgexit. If not under SVM, this would cause an #UD fault. In order to enable
the boot code to proceed anyway, implement a __vmgexit_safe() fixing up any
potential #UD in cooperation with the new ud_vmgexit_fixup_handler32() fault
handler.

Signed-off-by: Nicolai Stange <nstange@suse.de>