Stage2/boot: determine PTE C-bit more safely
Under SEV-SNP, the initial identiy mapping setup from the boot_stage2.rs
initialization code must have the C-bit set in the PTEs, or otherwise encrypted
pages' contents would be reinterpreted as plaintext, which would crash quickly.

As such, the boot code needs to determine the system specific C-bit positition
within the PTEs. Currently, this is implemented by examining the provided cpuid
table. However, the cpuid table is available only under SEV-SNP. If the user
mistakenly attempted to launch the SVSM under a non-SEV-SNP environment, it
would still be great if it would come and eventually terminate with a
meaningful error once the console becomes functional.

Make the boot_stage2.rs code to come up if SEV-SNP is not active. The different
possible scenarios are:
- SEV-SNP is active: the C-bit is needed. It can be obtained either from the
  cpuid table or from the host via the GHCB MSR protocol. A plain cpuid
  instruction cannot be used.
- SEV-ES is active: the C-bit is needed. Neither the cpuid table nor the
  plain cpuid instruction are available. The C-bit position must be obtained
  from the host by means of the GHCB MSR protocol.
- SEV is active: the C-bit is needed. The GHCB MSR protocol is not available,
  but a plain cpuid instruction can be used.
- SEV not active/not supported: C-bit does not exist. The SEV status MSR might
  not exist.

The main challenge is to figure out which of the above configuration is active
witout leveraging the cpuid instruction as this would cause immediate havoc
under SEV-ES or SEV-SNP. In theory, another vendor could have allocated
the SEV related MSR addresses to a completely different purpose, so even if
they exist, they cannot be relied upon without further verification. Vendor
confirmation via cpuid is not possible, so there's kind of a chicken-and-egg
problem here. Work around it by starting out with the assumption that the SEV
related MSRs, if existing, are indeed the SEV MSRs and run some consistency
checks on their contents and behaviour:
1.) The SEV status MSR should indicate that SEV-ES or SEV-SNP is active,
    otherwise we can resort to a plain cpuid instruction straight away.
2.) If the SEV status MSR indicates that SEV-ES or SEV-SNP is active, then
    the host should respond to GHCB MSR protocol information requests properly,
    supplying the C-bit alongside.

Implement this.

It's worth emphasizing that at this early point in the boot process, the C-bit
supplied by the untrusted HV is relied upon only for getting up and running
to the point of having a functional console. The C-bit position will later be
verified to match what is found in the more trustworthy SEV-SNP cpuid table.

Signed-off-by: Nicolai Stange <nstange@suse.de>