Stage2/boot: export vmgexit_safe() to post-boot Rust stage2
Export vmgexit_safe() to the stage2 part written in Rust so that the potentially
unsupported vmgexit instruction can be safely invoked from there.

Note that as the Rust part runs in long mode, this involves providing a 64 bit
variant of the #UD fixup handler relied upon by the wrapped __vmgexit_safe().

Signed-off-by: Nicolai Stange <nstange@suse.de>