Stage2: probe SEV_STATUS + GHCB MSR availability before using them
By the time the stage2 begins execution, the environment could in principle be
anything, including one with no SEV support at all. If the user happened to
mistakenly launch the SVSM under such an environment, it would be convenient if
a meaningful error message could still get printed to the console before
panicking.

To achieve this, the SEV_STATUS MSR must be accessed more carefully, as
non-existant MSR would #GP. Use the recovering rdmsr_safe() already exported
from the boot_stage2.rs code for first accesses to the SEV_STATUS MSR.

Lacking any cpuid information at this point, a successful read from the
SEV_STATUS MSR on its own does not confirm that SEV is supported -- vendors
could have allocated the MSR address to completely different purposes. Thus,
if the SEV-ES/-SNP related status flags are found to be set, it does not
necessarily mean that SEV-ES or -SNP is indeed enabled. But if they were, the
host should respond to the GHCB MSR protocol. So if the SEV_STATUS MSR suggests
that SEV-ES or SEV-SNP is active, go ahead and seek further confirmation by
issuing a GHCB MSR protocol information request and verifying the response.

Wrap all this in a new function, sev_ghcb_msr_available(). As a side-effect,
the GHCB MSR protocol information response will include the C-bit position.
Return it from sev_ghcb_msr_available() for use in a later patch.

Make stage2.rs to invoke the new sev_ghcb_msr_available() in place of the former
sev_es_enabled().

Signed-off-by: Nicolai Stange <nstange@suse.de>