Stage2: tentatively initialize paging early
If SEV-ES is enabled, but SEV-SNP is not, the cpuid page is not available
and paging_init() cannot obtain the required PTE C-bit position from it. As
a consequence, the GHCB and thus, the SVSMIOport based console depending on it,
would be disfunctional. As a consequence, any error reporting would not work in
this configuration.

However, the early sev_ghcb_msr_available() check would return the C-bit
position in either case, SEV-ES only or SEV-SNP enabled, as a side-effect.

In order to make SEV status error reporting work for SEV-ES only, use it to
initialize paging tentatively. Provide the new
mm::pagetable::paging_init_early() helper for this.

Coming from the HV, the C-bit position returned by sev_ghcb_msr_available()
cannot be trusted. If SEV-SNP is found active and the stage2 eventually
decides to continue and launch the svsm, make it, or rather the original
mm::pagetable::paging_early(), to verify the C-bit position found in the
more trustworthy cpuid page matches what had previously been obtained.

Signed-off-by: Nicolai Stange <nstange@suse.de>