mm/alloc: let MemoryRegion::phys_to_virt() accept any addr for identity mappings
Currently, there are three phys_to_virt() implementations around:
- the MemoryRegion::phys_to_virt(),
- one in the svsm bin crate, wrapping the former and
- another one in the stage2 bin crate accounting for the identity-mapping in
  effect at that stage and simply casting the supplied PhysAddr to a VirtAddr.

It's desirable to consolidate these three into one, in particular to avoid
dependencies from mm/pagetable.rs back to the respective bin crate.

However, as it currently stands, MemoryRegion::phys_to_virt() checks the
supplied PhysAddr to be within bounds of the registered region and errors out if
that's not the case.

That's going to be a problem when the pagetable walker is attempting to
translate the stage2's page table pages back, which are allocated statically at
link time and thus, outside of the registered memory region.

Prepare for consolidating the three phys_to_virt() implementations into one by
disabling the bounds check in MemoryRegion::phys_to_virt() for the special case
of an identity mapping.

Signed-off-by: Nicolai Stange <nstange@suse.de>