Stage2+SVSM: register CPUID_PAGE with the cpu::cpuid module
Currently, the cpu::cpuid modules constains link-time references to the static
CPUID_PAGE defined in the stage2 and svsm bin crates respectively, which hinders
making it part of a library crate.

Make the binary crates to register their CPUID_PAGE instance with the
cpu::cpuid module at runtime each and make the latter to store an internal
reference to it.

Signed-off-by: Nicolai Stange <nstange@suse.de>