Cargo manifest: disable doctests
In 'cargo test' environments, the SvsmAllocator must not be made the
global_allocator. However, the conditional setting of this attribute depends
on the 'test' configuration predicate being set in 'cargo test' environments.

This seems to not be working properly for documentation tests ([1]), and it's
similar for the alternative 'doctest' predicate ([2]).

As documentation tests currently don't play an important role for the project,
just disable them alltogether until the cargo community has settled on a
suitable alternative.

[1] https://github.com/rust-lang/rust/issues/45599
[2] https://github.com/rust-lang/rust/issues/67295

Signed-off-by: Nicolai Stange <nstange@suse.de>