lib: add nop test for verifying the test environment
Provide a trivial test for verifying that the 'cargo test' setup works as
expected.

Note that 'cargo test' needs to be made to build the library and tests for the
host, not the 'svsm-target' embedded target used for the stage2 and svsm binary
crates. This can be achieved with e.g.

  cargo test --target=x86_64-unknown-linux-gnu -Z build-std

Signed-off-by: Nicolai Stange <nstange@suse.de>