mm/alloc: determine buddy page solely based on the pfn
Currently, the code in MemoryRegion::compound_neighbor() finds the buddy for
a given comund page of specified pfn by first transforming the pfn to a
virtual address, flipping the bit corresponding to the given order and
transforming the resulting address back to a pfn.

This detour is not only unncessarily complex, it also breaks in case the memory
region's base address is not aligned to the maximum possible order, i.e.
to 2^5 * 4kB == 128 kB, because the bit flipping operation and the addition of
the base address to convert the pfn to a virtual page address are not
commutative.

On real HW this might not be a problem, because the memory region would perhaps
always be aligned by 128kB there (unverified), but for the regions to be setup
for the upcoming unit tests, this wouldn't be the case.

It is possible to determine a compound page's buddy directly from the pfn, i.e.
without that detour to virtual page addresses and back, so do that.

Signed-off-by: Nicolai Stange <nstange@suse.de>