sev/utils: refactor error handling
Introduce a new enum, SevSnpError, to hold possible error codes for
SEV-SNP instructions (RMPADJUST and PVALIDATE for now). This
obsoletes PValidateError.

We add an extra variant, FAIL_UNCHANGED, for cases where PVALIDATE
completed successfully but the RMP entry was not updated - the caller
can always choose to ignore this by inspecting the enum variant. This
variant is only returned if there were no other additional errors
(e.g. the instruction must return SUCCESS). This allows us to keep
the callers mostly unchanged since the RMP not being updated is
already being treated as an error in most places.

Each variant holds the error code returned by the instruction. For
FAIL_UNCHANGED, since we did not really get an error code from the
hardware, store the protocol-specific code that we want to return to
the guest.

While we are at it, introduce logging for the case something goes very
wrong and we get an unexpected return value, and panic afterwards
with unreachable!().

Also, remove rmpadjust_adjusted_error(), since it serves no purpose at
the moment, as no error codes higher than 6 can be returned from
rmp_adjust().

Signed-off-by: Carlos López <carlos.lopez@suse.com>