svsm/requests: refactor error handling
This follows up on a previous refactor of the error hanlding for
SEV-SNP instructions in sev/utils.

Introduce a new error type, SvsmError, to handle errors that may occur
when processing SVSM requests. The type has two variants,
RequestError (non-fatal) and FatalError. These variants are handled in
request_loop() - RequestError is reported to the guest, and
FatalError causes the SVSM to panic. For now, FatalError does not
contain any information as to what the source of the error was, but
this can be added in the future. For RequestError, an SvsmResultCode
is included.

SvsmResultCode maps the symbolic names given to errors in the SVSM
spec to a numeric code representing each. If a request fails, the
code is placed in VSMA.rax. SvsmError provides convenience methods to
construct a RequestError variant wrapping each of the possible
SvsmResultCode variants.

Refactor as well core request handlers to return a
Result<(), SvsmError>. Since the new VMSA.rax can be obtained from
this return value, refactor the handlers as well so that they take an
immutable reference to the VMSA instead of a mutable one, and assign
to RAX from the top of the request loop.

With this patch, error handling can be centralized to request_loop(),
the ? operator can be used more ergonomically, and additional
groundwork is laid for more granular logging and error handling.

Signed-off-by: Carlos López <carlos.lopez@suse.com>