svsm/requests: check that the guest exited with VMGEXIT
From the SVSM spec: "The SVSM is expected to examine VMSA.EXITCODE
after setting VMSA.EFER.SVME=0 to ensure that the guest is on the
expected VMGEXIT instruction boundary before proceeding. If the exit
code does not indicate exiting due to VMGEXIT, the SVSM should reset
VMSA.EFER.SVME=1 and take no further action before returning to the
guest."

Add a new enum for the guest exitcode, stored in VMSA.guest_exitcode.
Check that this value corresponds to a VMGEXIT before processing a
request.

Signed-off-by: Carlos López <carlos.lopez@suse.com>