fw_cfg: avoid O(n^2) behavior when validating flash regions
At the moment, reading a flash region is done through
FwCfg::get_flash_region(), which takes a region number, and iterates
through the firmware file to get the nth region. Since this is only
used to traverse all regions to validate them, iterate over them
linearly. Remove also a bounds check that we had to take in case an
invalid region number was specified.

Do this by implementing a new method for FwCfg that returns an
iterator over the flash regions. As before, if the 'etc/flash' file
is not found, or there are no regions, do nothing by returning an
empty iterator instead of an error. This obsoletes flash_region_count()
and get_flash_region().

While we are at it, simplify the validation loop in validate_flash()
using step_by().

Signed-off-by: Carlos López <carlos.lopez@suse.com>