SVSM/requests: Rework locking around core_create/delete_vcpu() calls
The locking currently used is prone to dead-locks. The main reason is
that VCPUs do unmap VMSA and CAA pages remotely. Implement a new sheme
were only the VCPU itself can update its mappings to VMSA and CAA
pages. This allows a much more relaxed locking-scheme, where remote
VCPUs only update the meta-data remotely and set a flag to notify the
target VCPU that mappings need to be updated.

Signed-off-by: Joerg Roedel <jroedel@suse.de>