SVSM/mm/guestmem: Fix GuestPtr::read() operations
There was a subtle bug in the GuestPtr()::read() code path which
caused the compiler (in release builds) to return uninitialized data
instead of the data read from the source pointer.

The reason was a compiler optimization around MaybeUninit where the
compiler lost track of whether the memory in the container was written
to. It assumed no write happened an just returned the uninitialized
data to the caller.

Fix this by working with raw pointers instead of VirtAddr in the
GuestPtr code paths and update the assembly constraints in the
do_movsb() function. The inout("rdi") constraint is especially
important, as it is telling the compiler the memory is actually
written inside the asm! statement. This helps the compiler to keep
track of the memory write and fixes the bug.

With these changes the SVSM built as release-target can successfully
boot a Linux guest.

Co-developed-by: Carlos López <carlos.lopez@suse.com>
Signed-off-by: Joerg Roedel <jroedel@suse.de>