SVSM/sev/utils: Detect exceptions correctly in rmp_adjust()
If the RMPADJUST instruction causes an exception, the RAX register will
still contain the virtual address and not the error code. This makes
error handling fail, so check for an exception first before examining
the return error code.

While there, also adjust the asm! constraints and print unknown error
codes in hex.

Signed-off-by: Joerg Roedel <jroedel@suse.de>