svsm/requests: fix race condition during SVSM_CORE_CREATE_VCPU
At the moment, when registering a VMSA for the new vCPU to be created,
a check is made in valid_phys_address() to avoid using a
guest-provided gPA that corresponds to another VMSA, via vmsa_exists().
However, vmsa_exists() is prone to race conditions, as it releases the
lock on the global registered VMSA list (PERCPU_VMSAS) on return. If
the guest issues another SVSM_CORE_CREATE_VCPU request from a different
CPU using the same gPA after the check has passed, we would end up with
two VMSAs using the same gPA.

Fix this by registering the VMSA earlier, and report an error from
register_guest_vmsa() if there is an attempt to use the same VMSA gPA
twice.

On top of that, register the new VMSA as unused, and only attempt to
unregister used VMSAs from SVSM_CORE_DELETE_CPU - this way, the VMSA
we are processing cannot be unregistered until we are done. On the
error path for SVSM_CORE_CREATE_VCPU, only attempt to unregister an
unused VMSA, and if the request is processed successfully, mark it as
used.

Signed-off-by: Carlos López <carlos.lopez@suse.com>