svsm/requests: fix race condition during SVSM_CORE_DELETE_CPU
At the moment, no checks are made on the guest-provided gPA for VMSA
unregistering. Check it via unregister_guest_vmsa() before issuing
any RMPADJUST instructions, ensuring that only a single CPU will
attempt to unregister a given VMSA.

While we are at it, clear EFER before issuing RMPADJUST, ensuring that
the VMSA is not executing.

Signed-off-by: Carlos López <carlos.lopez@suse.com>