cpu/percpu: refactor percpu VMSAs to a explicit type
Create a new type for the global per-cpu VMSA registry, and refactor
functions that operate on it as methods on this type. This follows
the pattern of other operations in in cpu/percpu.rs and is more
Rust-like. This way, we make it explicit to callers that they are
operating on a global (blocking) structure, and simplifies the imports
to operate on this structure.

Signed-off-by: Carlos López <carlos.lopez@suse.com>