kernel_launch: make the the kernel launch parameter's names more expressive
The stage2 -> SVSM kernel handover information kept in struct KernelLaunchInfo
includes three members describing the memory layout at the time of handover:

  - kernel_start: the beginning of the physical memory region usable by the
                  SVSM kernel, spanning the kernel image itself as well the
                  region's remainder to be used as a heap,
  - kernel_end:   the end of that aforementioned physical memory region,
  - virt_base:    the virtual address where the physical memory region had been
                  mapped contiguously by the stage2.

A future commit will make the stage2 to load the SVSM kernel from an ELF file
image and because the ELF "process" (virtual) memory layout is not necessarily
contiguous, some more details about the memory layout will need to get conveyed
to the SVSM kernel. At that point, the existing KernelLaunchInfo's parameters'
rather generic naming will easily give rise to confusion: it would e.g. not be
immediately clear at first sight that the memory region described by
->kernel_start and ->kernel_end does not refer to the loaded kernel ELF
"process" image layout, but to all of physical memory usable by the SVSM kernel,
including the heap. Likewise, "->virt_base" could perhaps be mistaken as the
ELF load base, which it is not.

Use a more explicit naming for the KernelLaunchInfo's members describing the
memory layout. While at it, drop the unused ->ghcb member.

This is a mere renaming patch, there is no change in functionality.

Signed-off-by: Nicolai Stange <nstange@suse.de>