Stage2, SVSM: pass heap area bounds from Stage2 to the SVSM kernel
With an upcoming change making the Stage2 loader to load the SVSM kernel from
an ELF file, the kernel will no longer be able to easily determine its heap area
memory layout by itself: the ELF format is more generic in that the different
memory regions can be layout non-contiguously in memory and as a result, the
virt->phys mapping could in principle become non-linear. It follows that all the
linear, 'offset' based virt->phys calculations in the SVSM kernel's
init_page_table() would become invalid in this situation, both for the
individual ELF regions as well as for the subsequent heap area not described by
the ELF metadata at all. The issue with the virt->phys mappings for the ELF
regions will be rectified later, when the kernel will be made to infer its own
layout from the ELF metadata, too. For the heap layout OTOH, the Stage2 will
need to convey the respective bounds to the kernel at handover at some point
and now is a good time to implement this, so that the upcoming ELF loading
logic in Stage2 can take its final form right away.

Signed-off-by: Nicolai Stange <nstange@suse.de>