sev/ghcb: support page state changes on regions with mixed 4kB/2MB pages
Currently, GHCB::page_state_change() is supposed to operate on regions composed
entirely of either 2MB or 4kB pages, as indicated by its 'huge' parameter.

However, the upcoming support for ELF file loading will need to handle regions
whose head and/or tail might be composed of a couple of 4kB pages with a
contiguous sequence of 2MB pages inbetween.

Make GHCB::page_state_change() support this scenario: if the 'huge' parameter
is set to true, check each page individually for whether or not
 - it is aligned to a 2MB boundary as well as
 - if a full 2MB page would still be contained in the region to process
and treat it either as a huge page if so or fall back to a 4kB page size
otherwise.

Signed-off-by: Nicolai Stange <nstange@suse.de>