mm/validate: introduce support for changing memory ranges' "validated" status
Currently, there are only primitives for changing the tracked "validated"
status of individual pages, either of 4kB or 2MB size. This requires a loop over
pages at callers whenever a whole range composed of multiple pages undergoes a
status change.

Implement helpers for setting and clearing a memory range pages' "validated"
status at once.

Re the naming, note that the existing pvalidate_range() would suggest names
for the new primitives that also have a "_range" suffix. However, this would
give rise to a name clash with the ValidBitmap::set_range() initialization
function. So rename the latter to ::set_region() to avoid that.

Signed-off-by: Nicolai Stange <nstange@suse.de>