Stage2: load the SVSM kernel from an ELF file
Currently, the SVSM kernel is distributed as a raw binary image more or less
just copied by the Stage2 over to the destination location in memory before
control gets transferred. In preparation for future enhancements like address
space randomization or pretty-printing of stacktrace entries, some more
metadata information about the process image will be needed.

Distribute the SVSM kernel as a regular ELF file instead of as a raw binary
and make the Stage2 loader load it to memory as described in the ELF file's
PT_LOAD program header entries.

Signed-off-by: Nicolai Stange <nstange@suse.de>