SVSM: merge ELF .<base>.* sections into their respective .<base>
Currently, the linked SVSM kernel ELF contains lots of sections of the
form .rodata.<some_unique_identifier>, similar for .data.

The linker script has it right for .bss and .text in that it specifies
a proper wildcard pattern on the respective input section: all input
sections of e.g. the form .text.* get merged into a single .text output
section.

Do the same for .data and .rodata.

Signed-off-by: Nicolai Stange <nstange@suse.de>