SVSM: link SVSM kernel as a PIE
In order to lay the grounds for future address space randomization, build
the SVSM kernel ELF as a position independent executable (PIE). Rust builds
it objects with relocation-model=pie per default, so all that is left is to
make the linker to emit the .dynamic information required by the relocating
ELF loader.

- amend build.rs to pass --pic-executable to the linker when linking the SVSM
- amend the SVSM linker script by corresponding ELF program header
  specifications, a PT_DYNAMIC entry in particular. As the specification
  of ELF program headers from linker scripts, if any, must be comprehensive,
  the needed PT_LOAD entries must be specified as well as a consequence.

Signed-off-by: Nicolai Stange <nstange@suse.de>