Stage2, SVSM: pass kernel ELF file memory location to the SVSM kernel
The SVSM kernel will need access to its ELF metadata in the future, for
e.g. inferring the needed memory mappings or for getting access to its
symbol table for pretty-printing stacktraces.

The Stage2 has it mapped in memory and at early SVSM kernel execution,
the Stage2 page tables are still in effect. Pass the ELF file memory
location from the Stage2 loader to the SVSM kernel at handover so
that the latter can grab all the information it needs from there before
switching to its own runtime page tables.

Signed-off-by: Nicolai Stange <nstange@suse.de>