svsm/requests: rename SvsmError to SvsmReqError
This prepares for a future commit where we add a generic error type,
SvsmError. We will also provide a way to convert from SvsmError to a
SvsmReqError that can be returned to the guest. This way, we can
define later what error code we want to return to the guest kernel
based on the SvsmError variant.

Signed-off-by: Carlos López <carlos.lopez@suse.com>