mm/alloc: fix optimized out inline assembly
Fix a subtle bug in allocate_zeroed_page() where inline assembly
setting a memory region to zero would get optimized out due to the
target address being set as an "in" parameter instead of "inout".
This would cause the compiler to assume that the contents of the
pointer were not changed, resulting in a crash during boot in release
builds release builds. This is similar to the bug fixed in
35b8a31aa2b96e86 ("SVSM/mm/guestmem: Fix GuestPtr::read()
operations").

Fix this by using zero_mem_region(), which casts the address to a raw
pointer before zeroing out the region, ensuring the operation does
not get opimized away.

Signed-off-by: Carlos López <carlos.lopez@suse.com>