Stage1: Pass launch data via stack to stage2
The launch information was passed via registers from stage1 to stage2.
This does not scale with the amount of information we need to pass up
the boot chain, so put all information at the bottom of the stage2
stack.

This included information about the kernel image in physical memory
and a potential kernel fs archive. The latter is not used yet, but
passed on to KernelLaunchInfo.

Signed-off-by: Joerg Roedel <jroedel@suse.de>

fix for launch stack