svsm: introduce the PhysAddr type
Replace the old PhysAddr typedef for an actual type backed by a usize.
Having a explicit type avoids potential mistakes such as passing a
virtual address to a function that operates on a physical address or
vice versa. It also allows a centralization of methods operating on
addresses, like alignment checks.

In preparation for a VirtAddr type, introduce an Address trait that
implements common functionality between virtual and physical
addresses. This avoids code repetition, while also allowing specific
actions to be implemented for a single type - an example of this
would be casting an address to a pointer, which only makes sense for
virtual addresses.

The abstration introduced by this type should be zero-cost at runtime,
as the structure is simply backed by an usize with no additional
fields. Conversion should be no different than converting to or from
the backing type.

Signed-off-by: Carlos López <carlos.lopez@suse.com>