utils/immut_after_init: implement runtime checks
Add debug-only runtime checks for ImmutAfterInitCell. This prevents
using an uninitialized cell, as well as unintended double
initialization bugs. As a consequence, ImmutAfterInitCell::init() and
reinit() no longer need to be unsafe, since checks are done before
accessing the inner type.

To keep track of the initialization state we use an AtomicBool with
relaxed ordering, for code simplicity reasons.

Signed-off-by: Carlos López <carlos.lopez@suse.com>