string: do not depend on nightly maybe_uninit features
As part of the effort to build COCONUT SVSM with stable Rust, remove
the dependency on the maybe_uninit_uninit_array and
maybe_uninit_array_assume_init nightly features, which are needed by
the FixedString structure. Replace the use of a MaybeUninit array
with an array of MaybeUninit types, and use raw pointers to retrieve
the initialized data.

Signed-off-by: Carlos López <carlos.lopez@suse.com>