svsm/edk2: Introduce OVMF as submodule and build
COCONUT-SVSM requires a build of OVMF that supports booting via an SVSM
module. A default OVMF X64 firmware does not include this capability.
Therefore, rather than instructing the developer how to clone and
prepare a build of EDK2/OVMF, this patch adds the correct EDK2
repository and branch as a submodule and builds it as part of the SVSM
makefile.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>