svsm: do not depend on nightly const_mut_refs feature
Luckily, mutable references depending on this feature can be converted
to immutable references or be removed alltogether.

The mutable pointer to a ConsoleWriter in Console can be const because
ConsoleWriter::put_byte() takes &self. On the other hand,
ImmutAfterInitRef::new_from_cell() requires using a raw pointer in a
const context, which will not work without the feature. Since it is
not in use, remove the method.

Signed-off-by: Carlos López <carlos.lopez@suse.com>