svsm/ovmf: Refactor OVMF guest code into module
COCONUT-SVSM currently exclusively launches guests under OVMF. However, we
should not assume this will always be the case. With this in mind, and
also to prepare for integrating the SVSM build with OVMF the current
OVMF specific code has been refactored into a module to abstract
specific steps away from svsm.rs.

If another guest firmware is required in the future, it will be quite
simple to create a trait from the ovmf_fw/OvmfFw structure to limit any
changes required to the main SVSM module.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>