svsm: Refactor ConsoleWriter into R/W Terminal
The serial port is used in SVSM for logging, requiring only output.
In order to support debugging via a gdb stub, it is required that the
serial port supports bidirectional communication. Therefore the
existing ConsoleWriter trait and SerialPort implementation have
been enhanced to include reading as well as writing serial data.

With this change, the name of the trait, ConsoleWriter is not
representative so this has been renamed to Terminal.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>