Add Github actions
Add automated actions to be executed on every push to the repository
and every pull request. At the moment, every step of the build
process is done except the final linking, because the runner image
(ubuntu-latest) is on Ubuntu 22.04, which does not have a new enough
binutils version. The actions run all the tests via make test, as
well as cargo fmt to keep the code format consistent.

Signed-off-by: Carlos López <carlos.lopez@suse.com>