svsm/debug: Add a gdbstub module for connection of a remote debugger
Adds a GDB stub module to coconut-svsm which can provide a remote debug
capability from the main SVSM entry point through to the launch of the
guest. It can remain active at this point and allows debugging of the
kernel request loop.

This patch introduces the GDB stub module but does not enable it within
SVSM. The stub is implemented behind a cargo 'feature' meaning that it
is completely excluded from release builds. In order to use the GDB
stub, the feature must be enabled and the GDB stub must be initialised
sometime after the SVSM main entry point.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>